+++ CHANGELOG.md
@@ -1,8 +1,20 @@
+# 8.5.0
+## ADDED
+- Custom marker color attribute
+- Custom icons for the hold action attribute can now be defined in the Editor preferences -> Misc.
+- Icon picker to the hold action attribute  #435
+- ENH_fnc_iconPicker, an simple Icon Picker
+
+## CHANGED
+- Fixed delete crew tool would not work on UAVs/UAV Turrets (Thanks hypoxia125) #433
+- Fixed that unit traits and attributes would conflict with ACE (Thanks Nomas-X for the detailed issue) #436
+- Fixed tooltips for the input list attribute control were swapped
+- Fixed minimap code would run 1 frame into eden preview disabling map indicators
+- Fixed that AI attributes would have no effect when entity is transfered to a headless client (Thanks KJW)
+- Minor script optimizations and formatting improvements
+
 # 8.4.0
 ## ADDED
-<!--
-- Custom marker color attribute
--->
 - Added localization for garrison combo entries
 - New Garrison tool
     - Garrison any building by simply dragging all selected entities onto it after pressing the "G" key
@@ -1110,7 +1122,7 @@ menu
 ## REMOVED
 - Import/Export scenario attributes. Scenario Attribute Manager takes care of that now
 - Removed on Event Code Attributes (Too complex for a simple attribute)
-- Colour Picker was removed. Use https://forums.bohemia.net/forums/topic/222104-color-picker/ instead
+- Color Picker was removed. Use https://forums.bohemia.net/forums/topic/222104-color-picker/ instead
 - Removed weapon sway and recoil attribute due to issues on dedicated server
 
 
@@ -1448,7 +1460,7 @@ menu
 
 ## ADDED
 - Attribute which adds all alive entities to allCurators regulary during a mission
-- Colour picker. Can be found in the debug tools category
+- Color picker. Can be found in the debug tools category
 - Attribute to disable team switch
 - Attribute to disable mine detection for AI units
 - Attribute to disable weapon aim for AI units